+++ components/custom-test/survey-test/graphs/AnyMeansTenureChoice.tsx
@@ -1,8 +1,8 @@
 import React from "react"
 import SurveyGraphCard from "@/components/custom-test/survey-test/SurveyGraphCard";
 import { Bar, BarChart, XAxis, YAxis, ResponsiveContainer, Cell, LabelList } from "recharts";
-import { useSurveyContext } from "../../../../app/survey/context";
-import { TENURE_COLORS } from "../../../../app/survey/constants"
+import { useSurveyContext } from "@context/surveyContext";
+import { TENURE_COLORS } from "../../../../lib/survey/constants"
 import { CustomLabelListContentProps } from "@/components/custom-test/dashboard-test/graphs-temp/shared";
 
 export interface CustomYTickProps {
