+++ components/custom-test/survey-test/graphs/HousingOutcomes.tsx
@@ -1,11 +1,11 @@
 import React, { useState } from "react"
-import { TickProps } from "@/app/survey/types";
+import { TickProps } from "@/lib/survey/types";
 import SurveyGraphCard from "@/components/custom-test/survey-test/SurveyGraphCard";
 import { Bar, BarChart, XAxis, YAxis, ResponsiveContainer } from "recharts";
-import { useSurveyContext } from "../../../../app/survey/context";
+import { useSurveyContext } from "@context/surveyContext";
 import SurveyTenureSelector from "../SurveyTenureSelector";
-import { TENURE_COLORS } from "../../../../app/survey/constants";
-import { getTopFive } from "../../../../app/survey/utils";
+import { TENURE_COLORS } from "../../../../lib/survey/constants";
+import { getTopFive } from "../../../../lib/survey/utils";
 
 export const HousingOutcomes = () => {
     const housingOutcomes = useSurveyContext().barOrPie.housingOutcomes;
