+++ components/custom-test/survey-test/graphs/SupportDevelopment.tsx
@@ -1,8 +1,8 @@
 import React from "react"
 import SurveyGraphCard from "@/components/custom-test/survey-test/SurveyGraphCard";
 import { PieChart, Pie, Legend, ResponsiveContainer, Cell } from "recharts";
-import { useSurveyContext } from "../../../../app/survey/context";
-import { SUPPORT_DEVELOPMENT_ORDER } from "../../../../app/survey/constants";
+import { useSurveyContext } from "@context/surveyContext";
+import { SUPPORT_DEVELOPMENT_ORDER } from "../../../../lib/survey/constants";
 
 export const SupportDevelopment = () => {
     let { supportDevelopment } = useSurveyContext().barOrPie;
